fix: Deleting a log library Does not clear log information
